+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/CredentialToBeIssued.kt
@@ -60,9 +60,7 @@ sealed class CredentialToBeIssued {
 /**
  * Represents a claim that shall be issued to the holder, i.e., serialized into the appropriate credential format.
  *
- * To issue nested structures in SD-JWT, pick one of two options:
- * - Pass a collection of [ClaimToBeIssued] in [value].
- * - Put dots `.` in [name], e.g. `address.region`
+ * To issue nested structures in SD-JWT, pass a collection of [ClaimToBeIssued] in [value].
  *
  * To issue an array of elements, use a collection of [ClaimToBeIssuedArrayElement] in [value].
  *